pyproject.toml
    [project]
    name = "inventory-management-system"
    version = "0.1.0"
    description = "Add your description here"
    readme = "README.md"
    requires-python = ">=3.12"
    dependencies = [
        "asgiref==3.8.1",
        "celery>=5.4.0",
        "certifi==2024.8.30",
        "crispy-bootstrap4==2024.10",
        "crispy-bootstrap5==2024.10",
        "distlib==0.3.9",
        "django==5.1.3",
        "django-celery-beat>=2.8.1",
        "django-celery-results>=2.6.0",
        "django-crispy-forms==2.3",
        "django-filter>=25.1",
        "django-phonenumber-field[phonenumbers]>=8.1.0",
        "djangorestframework>=3.16.1",
        "filelock==3.16.1",
        "flower>=2.0.1",
        "gunicorn==23.0.0",
        "httpx>=0.28.1",
        "packaging==24.2",
        "phonenumbers>=9.0.14",
        "pipenv==2024.4.0",
        "platformdirs==4.3.6",
        "psycopg>=3.2.10",
        "psycopg2-binary==2.9.9",
        "python-decouple==3.8",
        "python-dotenv==1.0.1",
        "redis>=6.4.0",
        "setuptools==75.5.0",
        "sqlparse==0.5.2",
        "tzdata==2024.2",
        "uvicorn[standard]>=0.36.0",
        "virtualenv==20.27.1",
        "whitenoise==6.8.2",
]
